ct: move placeholder type into level zero
the placeholder is managed by the level zero stm. ideally it shouldn't
leak out into the rest of the system, but we aren't quite there yet.
