+++ src/v/cloud_topics/batch_cache/BUILD
@@ -20,8 +20,8 @@ redpanda_cc_library(
         "//src/v/bytes:iobuf",
         "//src/v/cloud_topics:extent_meta",
         "//src/v/cloud_topics:logger",
-        "//src/v/cloud_topics:placeholder",
         "//src/v/cloud_topics:types",
+        "//src/v/cloud_topics/level_zero/stm:placeholder",
         "//src/v/config",
         "//src/v/model",
         "//src/v/ssx:future_util",
