+++ src/v/cloud_topics/frontend/BUILD
@@ -32,8 +32,8 @@ redpanda_cc_library(
         "//src/v/cloud_topics:app",
         "//src/v/cloud_topics:data_plane_api",
         "//src/v/cloud_topics:extent_meta",
-        "//src/v/cloud_topics:placeholder",
         "//src/v/cloud_topics/level_zero/reader",
+        "//src/v/cloud_topics/level_zero/stm:placeholder",
         "//src/v/model",
         "//src/v/raft",
         "//src/v/storage",
