+++ src/v/cloud_topics/level_zero/reader/BUILD
@@ -26,8 +26,8 @@ redpanda_cc_library(
         "//src/v/cloud_topics:extent_meta",
         "//src/v/cloud_topics:logger",
         "//src/v/cloud_topics:object_utils",
-        "//src/v/cloud_topics:placeholder",
         "//src/v/cloud_topics:types",
+        "//src/v/cloud_topics/level_zero/stm:placeholder",
         "//src/v/config",
         "//src/v/model",
         "//src/v/ssx:sformat",
@@ -57,9 +57,9 @@ redpanda_cc_library(
         "//src/v/cloud_io:remote",
         "//src/v/cloud_topics:extent_meta",
         "//src/v/cloud_topics:logger",
-        "//src/v/cloud_topics:placeholder",
         "//src/v/cloud_topics:types",
         "//src/v/cloud_topics/level_zero/reader:materialized_extent",
+        "//src/v/cloud_topics/level_zero/stm:placeholder",
         "//src/v/config",
         "//src/v/container:chunked_circular_buffer",
         "//src/v/model",
@@ -85,12 +85,12 @@ redpanda_cc_library(
         "//src/v/cloud_io:basic_cache_service_api",
         "//src/v/cloud_io:remote",
         "//src/v/cloud_topics:logger",
-        "//src/v/cloud_topics:placeholder",
         "//src/v/cloud_topics:types",
         "//src/v/cloud_topics/level_zero/pipeline:event_filter",
         "//src/v/cloud_topics/level_zero/pipeline:read_pipeline",
         "//src/v/cloud_topics/level_zero/pipeline:read_request",
         "//src/v/cloud_topics/level_zero/reader:materialized_extent_reader",
+        "//src/v/cloud_topics/level_zero/stm:placeholder",
         "//src/v/model",
         "//src/v/ssx:future_util",
         "//src/v/storage",
@@ -111,7 +111,7 @@ redpanda_cc_library(
         "//src/v/cloud_topics:app",
         "//src/v/cloud_topics:data_plane_api",
         "//src/v/cloud_topics:logger",
-        "//src/v/cloud_topics:placeholder",
+        "//src/v/cloud_topics/level_zero/stm:placeholder",
         "//src/v/cluster",
         "//src/v/config",
     ],
