+++ src/v/cloud_topics/level_zero/reader/tests/BUILD
@@ -22,9 +22,9 @@ redpanda_test_cc_library(
         "//src/v/cloud_storage_clients",
         "//src/v/cloud_topics:extent_meta",
         "//src/v/cloud_topics:object_utils",
-        "//src/v/cloud_topics:placeholder",
         "//src/v/cloud_topics:types",
         "//src/v/cloud_topics/level_zero/reader:materialized_extent",
+        "//src/v/cloud_topics/level_zero/stm:placeholder",
         "//src/v/container:fragmented_vector",
         "//src/v/model",
         "//src/v/model/tests:random",
@@ -47,10 +47,10 @@ redpanda_cc_gtest(
     ],
     deps = [
         "//src/v/cloud_topics:extent_meta",
-        "//src/v/cloud_topics:placeholder",
         "//src/v/cloud_topics:types",
         "//src/v/cloud_topics/level_zero/reader:materialized_extent",
         "//src/v/cloud_topics/level_zero/reader/tests:materialized_extent_fixture",
+        "//src/v/cloud_topics/level_zero/stm:placeholder",
         "//src/v/container:fragmented_vector",
         "//src/v/model",
         "//src/v/ssx:sformat",
