+++ src/v/cloud_topics/level_zero/stm/placeholder.h
